Working on the unified schematic for the laser rangefinder.
The support circuitry for the at90usb1286 is being drawn from the
Teensy++2.0 released schematic.  Notable points are the direct usb
connection, the external ceramic 16MHz oscillator, and the external
reset and bootloader buttons.

According to the 1286 documentation, the chip ships with a USB DFU
class bootloader enabled by default.  Pulling the HWB pin low during a
reset (say, with a button) enables the bootloader, which presents a
DFU class to the host pc over the USB connection.  The microcontroller
may then be programmed with dfu-programmer or dfu-util.  In addition,
the user code may shut down all peripherals and then jump to the
bootloader (default address 0xf000), enabling software-controlled
programming.

As an aside, why is it that any time programmers sit down to write EDA
applications, the result is a steaming pile of crap?  KiCad is up
there among the worst, although they're all terrible.  I need to make
a LaTeX for schematic capture and board layout.